faulted lstm code with TypeError
Tensors in list passed to 'values' of 'ConcatV2' Op have types [int32, float32] that don't all match.

+++ src/models/features.py
@@ -82,7 +82,7 @@ class Features():
                 fDict = self.get_features()
                 v = []
                 for i in fDict:
-                        v.append(int(fDict[i]))
+                        v.append(float(fDict[i]))
 
                 return v
